CommanderAI/CommanderAI/requirements.txt
discord.py==2.4.1
requests
python-dotenv